+++ src/Bunny.h
@@ -10,14 +10,14 @@ public:
     Bunny(std::string a_name
         , std::string a_sex
         , std::string a_color
-        , int a_age
+        , int8_t a_age
         , bool a_mutant);
     
     // getters
     std::string getName()     const;
     std::string getSex()      const;
     std::string getColor()    const;
-    int         getAge()      const;
+    int8_t      getAge()      const;
     bool        getIsMutant() const;
 
 
@@ -31,7 +31,7 @@ private:
     std::string m_name;
     std::string m_sex;
     std::string m_color;
-    int         m_age;
+    int8_t      m_age;
     bool        m_isRadioactiveVampireMutant;
 };
